Move latex.* files
